Melhorado cadastro de colaboradores organizado botoes adicionar e adicionado mascara para email e telefone.

+++ app/Http/Controllers/Web/ClinicController.php
@@ -3,13 +3,15 @@
 namespace App\Http\Controllers\Web;
 
 use App\Clinic;
-use App\Collaborator;
 use App\Address;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\JsonResponse;
+use App\Collaborator;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Collection;
+use App\Http\Controllers\Controller;
+use App\Http\Requests\Clinic\storeClinicRequest;
+use App\Http\Requests\Clinic\updateClinicRequest;
 
 class ClinicController extends Controller
 {
@@ -30,54 +32,57 @@ class ClinicController extends Controller
      */
     public function create(Request $request)
     {
-        return view('clinic.form_create', compact('data'));
+        return view('clinic.form');
     }
-    public function store(Request $request){
-        $clinic = new Clinic();
-        $clinic->name = $request->name;
-        $clinic->email = $request->email;
-        $clinic->phone = $request->phone;
-        $clinic->save();
+    public function store(storeClinicRequest $request){
+        Clinic::create([
+            'name' => $request->name,
+            'email' => $request->email,
+            'phone' => $request->phone,
+            'address' => $request->address
+        ]);
 
-        return redirect()->route('clinic.index');
+        return redirect()->route('clinic.index')->with('status', 'Clínica cadastrada com sucesso.');
     }
     /**
      * edit the Clinic
      * @var Request $request
      * @return JsonResponse
      */
-    public function edit(Request $request, $id)
+    public function edit(Clinic $clinic)
     {
-
-        $clinic = Clinic::find($id);
-        return view('clinic.form_update', $clinic);
+        return view('clinic.form')->with('clinic', $clinic);
     }
     /**
      * Updates the Clinic
      * @var Request $request
      * @return JsonResponse
      */
-    public function update(Request $request, $id)
+    public function update(updateClinicRequest $request, $id)
     {
         $clinic = Clinic::find($id);
+
         $clinic->name = $request->name;
         $clinic->email = $request->email;
         $clinic->phone = $request->phone;
+        $clinic->address = $request->address;
 
         $clinic->save();
 
-        return redirect()->route('clinic.index');
+        return redirect()->route('clinic.index')->with('status', 'Clínica atualizada com sucesso.');
     }
     /**
      * Removes a clinic by it's id
      *
      * @return Collection
      */
-    public function delete($id)
+    public function destroy(Clinic $clinic)
     {
-        $clinic = Clinic::find($id);
+        if($clinic->appointments->count() > 0){
+            return redirect()->route('clinic.index')->with('status-alert','Clínica não pode ser deletada por estar vinculada a pelo menos um agendamento.');
+        }
         $clinic->delete();
         $clinic->deleted = true;
-        return new JsonResponse($clinic);
+        return redirect()->route('clinic.index')->with('status', 'Clínica deletada com sucesso.');
     }
 }
